requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==5.2.1
celery==5.4.0
python-dotenv==1.0.1
aiosmtplib==3.0.2
email-validator==2.2.0
